% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{EM_lucid}
\alias{EM_lucid}
\title{EM algorithm for LUCID with multi-omics data}
\usage{
EM_lucid(
  G,
  Z,
  Y,
  K,
  modelNames = rep("VVV", length(K)),
  useY = TRUE,
  init_par = NULL,
  tol = 0.001,
  family = c("gaussian", "binomial"),
  max_itr = 1000,
  seed = 123
)
}
\arguments{
\item{G}{a matrix}

\item{Z}{a list, each element is a matrix}

\item{Y}{a vector}

\item{K}{a vector, each element is a integer, representing number of latent
clusters corresponds to each element in Z}

\item{modelNames}{a vector of string specifies the geometric feature of omics
data. See more in ?mclust::mclustModelNames}

\item{useY}{logical, if TRUE, LUCID will use the information of Y to estimate
the latent cluster}

\item{init_par}{To be continued. An interface to initialize EM algorithm}
}
\value{
A list contains the object below:
\enumerate{
\item res_Beta: estimation for G->X associations
\item res_Mu_Sigma: estimation for X->Z associations
\item res_Delta: estimation for X->Y associations
\item loglik: log likelihood of LUCID model
\item z: posterior cluster assignment for each observation
\item K: number of latent clusters
\item N: number of observations
}
}
\description{
EM algorithm for LUCID with multi-omics data
}
