% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_lucid}
\alias{predict_lucid}
\title{Prediction function for LUCID}
\usage{
predict_lucid(model, G = NULL, Z = NULL, Y = NULL)
}
\arguments{
\item{model}{A LUCID model returned by EM_lucid}

\item{G}{matrix}

\item{Z}{a list, each element is a matrix}

\item{Y}{vector}
}
\value{
a list contains
\enumerate{
\item predicted cluster assignment for each omic layer
\item posterior inclusion probability
\item predicted outcome
}
}
\description{
Prediction function for LUCID
}
