% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{reorder_lucid}
\alias{reorder_lucid}
\title{function to reorder all model parameters}
\usage{
reorder_lucid(model)
}
\arguments{
\item{model}{A model returned by EM_lucid}
}
\value{
A LUCID model reordered by effect size of outcome
}
\description{
function to reorder all model parameters
}
